+++ includes/debug/zuplus-debug.php
@@ -6,7 +6,6 @@ include_once('debug-bar.php');
 
 class ZU_Debug extends zuplus_Addon {
 
-	private $dlog;
 	private $alog;
 	private $profiler;
 	
@@ -21,6 +20,18 @@ class ZU_Debug extends zuplus_Addon {
 	private $location_priority;
 	private $abs_path;
 	private $content_path;
+
+    public static $_defaults = [
+		'ajax_log'					=>	false,
+		'profiler'					=>	false,
+		'debug_js'					=>	false,
+		'debug_cache'			=>	false,
+		'debug_bar'				=>	true,
+		'debug_backtrace'		=>	false,
+		'write_to_file'			=>	false,
+		'output_html'			=>	true,
+		'beautify_html'			=>	true,
+    ];
 
 	protected static $ignore_class = [
 		'wpdb'           		=> true,
@@ -62,7 +73,6 @@ class ZU_Debug extends zuplus_Addon {
 	
 	protected function construct_more() {
 		
-		$this->dlog = $this->check_option('debug_log');
 		$this->alog = $this->check_option('ajax_log');		
 		$this->profiler = $this->check_option('profiler');	
 		$this->use_backtrace = $this->check_option('debug_backtrace');	
@@ -272,8 +282,6 @@ class ZU_Debug extends zuplus_Addon {
 	
 	public function write_trace($msg, $full_trace = false, $ignore_args = true) {
 		
-		if(!$this->dlog) return;
-		
 		$trace = $full_trace ? $this->get_full_backtrace($ignore_args) : $this->get_backtrace();
 		unset($trace['args'], $trace['type']);
 
@@ -289,8 +297,6 @@ class ZU_Debug extends zuplus_Addon {
 	}
 	
 	public function write_log($msg, $var = 'novar', $bt = false, $save_debug_bar = true, $forced_write_file = false) {
-		
-		if(!$this->dlog) return;
 		
 		$trace = ($this->use_backtrace || $bt)  ? $this->get_backtrace() : [];
 		$f =  $this->log_location();
@@ -369,42 +375,87 @@ class ZU_Debug extends zuplus_Addon {
 	public function get_request_ip() { return ZU_DebugBar::get_server_value('REMOTE_ADDR'); }
 	public function get_request_ajax() { return ZU_DebugBar::get_server_value('AJAX'); }
 	public function get_request_refer() { return ZU_DebugBar::get_server_value('HTTP_REFERER'); }
+	
+	// Admin functions -----------------------------------------------------------]
+	
+	public function print_log_location($post) {
+
+		$form = $this->get_form();
+		if(empty($form)) return;
+
+		$form->add_value('log_location', $this->log_location());
+		$form->text('log_location', 'Log Location', '', true);
+	
+		echo $form->fields('It can be changed with the function <span>_dbug_change_log_location()</span>.');
+	}
+	
+	public function print_stats() {
+	
+		$stats = $this->log_stats();
+		return sprintf('<p>Log size: <span>%1$s</span></p><p>Log priority: <span>%2$s</span></p>', $stats['size'], $stats['priority']);
+	}
+
+	public function print_debug_options($post) {
+
+		$form = $this->get_form();
+		if(empty($form)) return;
+
+		$form->checkbox('debug_bar', 'Use Debug Bar', 'Works only if <span>Query Monitor</span> is activated.');
+		$form->checkbox('debug_js', 'Activate JS Debug info', 'Adds class <span>debug</span> to BODY and displays debug info for responsive elements.');
+
+		$form->checkbox('debug_cache', 'Debug Caching', 'If checked, all calls to cache functions will be logged.');
+		$form->checkbox('debug_backtrace', 'Always Include Backtrace', 'In some cases, this can <span>greatly slow down</span> the loading of the page and even lead to a fatal error.');
+		$form->checkbox('write_to_file', 'Write log to file', 'If unchecked, only the information for <span>Debug Bar</span> will be saved.');
+		$form->checkbox('output_html', 'Display HTML entities in Debug Bar', 'If checked, all characters which have HTML character entity equivalents are translated into these entities.');
+		$form->checkbox('beautify_html', 'Beautify HTML in output', 'If unchecked, all HTML values will be saved without any modifications. Otherwise HTML beautifier will be used.');
+
+		$form->checkbox('ajax_log', 'Activate AJAX Logging', 'You should make <span>AJAX calls</span> from your JS.');
+		$form->checkbox('profiler', 'Activate Profiler', 'You should call <span>_profiler_flag()</span> at each point of interest, passing a descriptive string.');
+		
+		echo $form->fields('Debug Mode Settings.');
+	}	
 } 
 
 // Functions for use in code --------------------------------------------------]
 
 if(!function_exists('zu_write_log')) {
 	function zu_write_log($msg, $var = 'novar') {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_log($msg, $var);
 	}
 }
 
 if(!function_exists('_dbug_use_var_dump')) {
 	function _dbug_use_var_dump($dump = true) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->use_var_dump($dump);
 	}
 }
 
 if(!function_exists('_dbug_change_log_location')) {
 	function _dbug_change_log_location($path, $priority = 1) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->change_log_location($path, $priority);
 	}
 }
 
 if(!function_exists('_dbug_log')) {
 	function _dbug_log($msg, $var = 'novar', $bt = false) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_log($msg, $var, $bt);
 	}
 }
 
 if(!function_exists('_dbug_log_only')) {
 	function _dbug_log_only($msg, $var = 'novar', $bt = false) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_log_no_debug_bar($msg, $var, $bt);
 	}
 }
 
 if(!function_exists('_dbug_dump')) {	// Use this function to output structured information. Arrays and objects are explored recursively with values indented to show structure. 
 	function _dbug_dump($msg, $var = 'novar', $bt = false) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->use_var_dump(true);
 		zuplus_instance()->dbug->write_log($msg, $var, $bt);
 		zuplus_instance()->dbug->use_var_dump(false);
@@ -413,30 +464,35 @@ if(!function_exists('_dbug_dump')) {	// Use this function to output structured i
 
 if(!function_exists('_dbug_trace')) {
 	function _dbug_trace($msg, $full_trace = false) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_trace($msg, $full_trace);
 	}
 }
 
 if(!function_exists('_dbug_log_if')) {
 	function _dbug_log_if($condition, $msg, $var = 'novar', $bt = false) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_log_if($condition, $msg, $var, $bt);
 	}
 }
 
 if(!function_exists('_profiler_flag')) {
 	function _profiler_flag($flag_name) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->profiler_flag($flag_name);
 	}
 }
 
 if(!function_exists('_tbug_log')) {
 	function _tbug_log($log_name, $log_value) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->save_log($log_name, $log_value);
 	}
 }
 
 if(!function_exists('_ajax_log')) {
 	function _ajax_log($data) {
+		if(zuplus_nodebug()) return;
 		zuplus_instance()->dbug->write_ajax_log($data);
 	}
 }