+++ zu-plus.php
@@ -4,7 +4,7 @@ Plugin Name: ZU+
 Plugin URI: https://dmitryrudakov.ru/plugins/
 GitHub Plugin URI: https://github.com/picasso/zu-plus
 Description: This plugin encompasses ZU framework functionality.
-Version: 1.3.8
+Version: 1.4.1
 Author: Dmitry Rudakov
 Author URI: https://dmitryrudakov.ru/about/
 Text Domain: zu-plugin
@@ -32,7 +32,7 @@ Domain Path: /lang/
 
 // Prohibit direct script loading
 defined('ABSPATH') || die('No direct script access allowed!');
-define('ZUPLUS_VERSION', '1.3.8');
+define('ZUPLUS_VERSION', '1.4.1');
 define('ZUPLUS_NAME', 'ZU+');
 define('__ZUPLUS_ROOT__', plugin_dir_path(__FILE__)); 
 define('__ZUPLUS_FILE__', __FILE__); 
@@ -49,10 +49,14 @@ define('QM_HIDE_SELF', true);		// Hides the internal actions of Query Monitor in
 // define('GITHUB_UPDATER_OVERRIDE_DOT_ORG', true);
 //
 
+// Uncomment to debug plugin constructor (which is earlier than creation of ZU Debug)
+// include_once('/nas/content/live/dmitryrudakov/wp-content/plugins/zu-plus/includes/debug/sys-debug.php');
+// _sdbug_location_plugins('zu-plus'); 
+
 class ZU_Plugin extends zuplus_Plugin {
 
-	public $dbug;
-	public $cnotice;
+	public $dbug = null;
+	public $cnotice = null;
 
 	protected function extend_config() {
 		return  [
@@ -67,24 +71,36 @@ class ZU_Plugin extends zuplus_Plugin {
 	
 	protected function construct_more() {
 		
-		$this->dbug = new ZU_Debug($this->config_addon()); 
-		zu()->set_debug_cache($this->check_option('debug_cache'));
-
+		if($this->check_option('debug_mode')) {
+			$this->dbug = $this->create_addon(ZU_Debug::class); 
+			zu()->set_debug_cache($this->check_option('debug_cache'));
+		}
+		
 		if($this->check_option('cookie_notice')) {
-			$this->cnotice = new ZU_CookieNotice($this->config_addon());
+			$this->cnotice = $this->create_addon(ZU_CookieNotice::class);
 		}
 		
 		add_filter('zu_update_defaults', function($zu_defaults, $key = null) {
 			
 			if(empty($key)) {
 				$options = $this->options();
-				zu()->set_option($zu_defaults, 'refresh_mode', $this->check_option('debug_log'));
-				zu()->set_option($zu_defaults, 'cookie_notice', $this->check_option('cookie_notice'));
+				zu()->set_option($zu_defaults, 'refresh_mode', $this->check_option('debug_mode'));
+				zu()->set_option($zu_defaults, 'ajax_log', $this->check_option('ajax_log'));
+				zu()->set_option($zu_defaults, 'debug_mode', $this->check_option('debug_js'));
+				
+				if($this->check_option('cookie_notice')) {
+					zu()->set_option($zu_defaults, 'cookie_notice', true);
+					zu()->set_option($zu_defaults, 'cookie_options', $this->option_value('cookie_options', []));
+				}
 			}
 			
 			return $zu_defaults;
 			
 		}, 10, 2);
+	}
+	
+	public function is_debug() {
+		return empty($this->dbug) ? false : true;
 	}
 }
 
@@ -95,8 +111,12 @@ class ZU_Admin extends zuplus_Admin {
 	protected function construct_more() {
 		
 		if($this->check_option('dup_page')) {
-			$this->duppage = new ZU_DuplicatePage($this->config_addon());
+			$this->duppage = $this->create_addon(ZU_DuplicatePage::class);
 		}
+	}
+	
+	private function  is_debug() {
+		return $this->plugin->is_debug();
 	}
 	
 	//
@@ -175,22 +195,21 @@ class ZU_Admin extends zuplus_Admin {
 	}
 
 	protected function options_defaults() { 
-		$zu_defaults = [
-			'debug_log' 				=>	true,
-			'ajax_log'					=>	false,
-			'profiler'					=>	false,
-			'debug_backtrace'		=>	false,
-			'write_to_file'			=>	false,
-			'beautify_html'			=>	true,
-			'output_html'			=>	true,
-			'debug_bar'				=>	true,
-			'debug_cache'			=>	false,
-			
-			'dup_page'				=>	false,
-			'cookie_notice'			=>	false,
-		]; 
+		//  debug only with _sdbug
 		
-		return array_merge($zu_defaults, ZU_DuplicatePage::dup_defaults());
+		$zu_defaults = [
+			'debug_mode'			=>	false,
+			'cookie_notice'			=>	false,
+			'dup_page'				=>	false,
+		]; 
+
+		$zu_addons = [
+			'debug_mode'			=>	ZU_Debug::class,
+			'cookie_notice'			=>	ZU_CookieNotice::class,
+			'dup_page'				=>	ZU_DuplicatePage::class,
+		];
+
+		return $this->preprocess_defaults($zu_defaults, $zu_addons);
 	}
 
 	public function validate_options($input) {
@@ -212,63 +231,67 @@ class ZU_Admin extends zuplus_Admin {
 	
 	public function meta_boxes_more($settings_page, $no_default_boxes) {
 
-		// Add button to clear logs -------------------------------------------------]
-		
-		add_filter($this->prefix.'_print_debug_buttons', function() {
-			return $this->form->button_link('zuplus_clear_log', __('Clear Debug Log', 'zu-plugin'), 'trash', 'blue');
-		});
-
 		// Custom Boxes -------------------------------------------------------------]
 		
-		$this->form->add_meta_box('log', __('Actual Log Location', 'zu-plugin'), [$this, 'print_log_location']);
-		$this->form->add_meta_box('duplicate_menu', __('Duplicate Menu', 'zu-plugin'), [$this, 'print_duplicate_menu']);
+		$this->form->add_meta_box('all_actions', __('Theme Actions', 'zu-plugin'), [$this, 'print_all_actions']);
+
+		if($this->check_option('debug_mode')) {
+
+			$this->form->add_meta_box('debug_options', __('Debug Options', 'zu-plugin'), [$this->plugin->dbug, 'print_debug_options'], 'advanced', 'high');
+			$this->form->add_meta_box('debug_log_location', __('Actual Log Location', 'zu-plugin'), [$this->plugin->dbug, 'print_log_location'], 'advanced', 'low');
+	
+			// Add button to clear logs ------------------------------------------------]
+		
+			add_filter($this->prefix.'_print_debug_buttons', function() {
+				return $this->form->button_link('zuplus_clear_log', __('Clear Debug Log', 'zu-plugin'), 'trash', 'blue');
+			});
+		}
 
 		if($this->check_option('dup_page')) {
-			$this->form->add_meta_box('duplicate_page', __('Duplicate Page', 'zu-plugin'), [$this, 'print_duplicate_page']);
+			$this->form->add_meta_box('duplicate_page', __('Duplicate Page', 'zu-plugin'), [$this->duppage, 'print_duplicate_page'], 'advanced');
+			$this->form->add_meta_box('duplicate_menu', __('Duplicate Menu', 'zu-plugin'), [$this, 'print_duplicate_menu']);
 		}
 
 		if($this->check_option('cookie_notice')) {
-			$this->form->add_meta_box('cookie_notice', __('Cookie Notice', 'zu-plugin'), [$this, 'print_cookie_notice']);
+			$this->form->add_meta_box('cookie_notice', __('Cookie Notice', 'zu-plugin'), [$this->plugin->cnotice, 'print_cookie_metabox'], 'advanced');
 		}
-
-		$this->form->add_meta_box('all_actions', __('Theme Actions', 'zu-plugin'), [$this, 'print_all_actions'], 'advanced');
 	}
 
 	public function status_callback() {
-		$stats = $this->plugin->dbug->log_stats();
-		return sprintf('<p>Log size: <span>%1$s</span></p><p>Log priority: <span>%2$s</span></p>', $stats['size'], $stats['priority']);		
+		
+		return $this->is_debug() ? $this->plugin->dbug->print_stats() : '';
 	}
 
 	public function print_options($post) {
 
-		$this->form->checkbox('ajax_log', 'Activate AJAX Logging', 'You should make <span>AJAX calls</span> from your JS.');
-		$this->form->checkbox('profiler', 'Activate Profiler', 'You should call <span>_profiler_flag()</span> at each point of interest, passing a descriptive string.');
-
-		$this->form->checkbox('debug_log', 'Activate Debug Logging', 'All calls to <span>_dbug_log()</span> functions will written to logfile.');
-
-		$this->form->checkbox('debug_cache', 'Debug Caching', 'If checked, all calls to cache functions will be logged.');
-		$this->form->checkbox('debug_bar', 'Use Debug Bar', 'Works only if <span>Query Monitor</span> is activated.');
-		$this->form->checkbox('debug_backtrace', 'Always Include Backtrace', 'In some cases, this can <span>greatly slow down</span> the loading of the page and even lead to a fatal error.');
-		$this->form->checkbox('write_to_file', 'Write log to file', 'If unchecked, only the information for <span>Debug Bar</span> will be saved.');
-		$this->form->checkbox('output_html', 'Display HTML entities in Debug Bar', 'If checked, all characters which have HTML character entity equivalents are translated into these entities.');
-		$this->form->checkbox('beautify_html', 'Beautify HTML in output', 'If unchecked, all HTML values will be saved without any modifications. Otherwise HTML beautifier will be used.');
-		
-		$this->form->checkbox('dup_page', 'Activate Duplicate Page', 'Allows duplicate Posts, Pages and Custom Posts using single click.');
-
+		$this->form->checkbox('debug_mode', 'Activate Debug Mode', 'All debug functions like <span>_dbug_*()</span> will be activated. Otherwise all calls will be muted.');
+		$this->form->checkbox('dup_page', 'Activate Duplicate Page & Menu', 'Allows duplicate Menu, Posts, Pages and Custom Posts using single click.');
 		$this->form->checkbox('cookie_notice', 'Activate Cookie Notice', 'Allows you to inform users that the site uses cookies and to comply with the EU GDPR regulations.');
 	
 		echo $this->form->fields('The plugin encompasses ZU framework functionality.');
 		echo $this->form->print_save_mobile();
 	}
 
-	public function print_log_location($post) {
+	public function print_all_actions($post) {
 
-		$this->form->add_value('log_location', $this->plugin->dbug->log_location());
-		$this->form->text('log_location', 'Log Location', '', true);
+		$this->form->button_link_with_help('zuplus_reset_cached', 
+			__('Reset All Cached Shortcodes', 'zu-plugin'), 
+			'dismiss', 
+			'magenta', 
+			'Clear all cached data referenced to shortcodes (<strong>gallery</strong> and <strong>select</strong>). Needs if something went wrong.'
+		);
 
-		echo $this->form->fields('It can be changed with the function <span>_dbug_change_log_location()</span>.');
+		$this->form->button_link_with_help('zuplus_revoke_cookie', 
+			__('Revoke Cookie', 'zu-plugin'), 
+			'hidden', 
+			'gold', 
+			'Set "expires" value on cookie for 1970 which leads to cookie <strong>deleting</strong>. Needs for debugging only.',
+			ZU_CookieNotice::cookie_name()
+		);
+
+		echo $this->form->fields('Actions available for ZU Theme.', 'zuplus_reset_cached', true); // second argument -> data-ajaxrel : used in js to serialize form
 	}
-
+	
 	public function print_duplicate_menu($post) {
 
         $nav_menus = wp_get_nav_menus();
@@ -285,62 +308,21 @@ class ZU_Admin extends zuplus_Admin {
 			$this->form->add_value('new_menu', '?');
             $this->form->text('new_menu', 'And call it', 'The name will be assigned to duplicated menu.');
         
-			$this->form->button_link('zuplus_duplicate_menu', __('Duplicate', 'zu-plugin'), 'images-alt2', 'red', true, false);
+			$this->form->button_link_with_help('zuplus_duplicate_menu', 
+				__('Duplicate Menu', 'zu-plugin'), 
+				'plus-alt', 
+				'magenta', 
+				'Select menu from dropbox and then give it a new name. A copy of selected menu will be created.'
+			);
 		}
 		
 		echo $this->form->fields($desc, 'zuplus_duplicate_menu', true); // second argument -> data-ajaxrel : used in js to serialize form
 	}
-
-	public function print_duplicate_page($post) {
-
-		$this->form->select(
-			'dup_status', 
-			'Duplicate Post Status:', 
-			$this->duppage->status_values(), 
-			'Select any post status you want to assign for duplicate post.'
-		);
-
-		$this->form->select(
-			'dup_redirect', 
-			'Redirect to after click on link:', 
-			$this->duppage->redirect_values(), 
-			'Select any post redirection after click on <strong>"Duplicate This"</strong> link.'
-		);
-		
-        $this->form->text('dup_suffix', 'Duplicate Post Suffix', 'Add a suffix for duplicate post as Copy, Clone etc. It will show after title.');
-		
-		echo $this->form->fields('Duplicate Page Settings.');
-	}
-
-	public function print_cookie_notice($post) {
-
-        $this->form->text('cookie_message', 'Notice Message', 'Enter the cookie notice message.');
-		echo $this->form->fields('Cookie Notice Settings.');
-	}
-
-/*
-__('Enter the cookie notice message.', 'cookie-notice')
-__('The text of the option to accept the usage of the cookies and make the notification disappear.', 'cookie-notice')
-__('Synchronize with WordPress Privacy Policy page.', 'cookie-notice')
-__('Enter the full URL starting with http(s)://', 'cookie-notice')
-__('Select the privacy policy link target.', 'cookie-notice')
-*/
-	
-	public function print_all_actions($post) {
-
-		$this->form->button_link_with_help('zuplus_reset_cached', 
-			__('Reset All Cached Shortcodes', 'zu-plugin'), 
-			'dismiss', 
-			'magenta', 
-			'Clear all cached data referenced to shortcodes (<strong>gallery</strong> and <strong>select</strong>). Needs if something went wrong.'
-		);
-
-		echo $this->form->fields('Actions available for ZU Theme.', 'zuplus_reset_cached', true); // second argument -> data-ajaxrel : used in js to serialize form
-	}
 	
 	public function ajax_more($option_name, $ajax_value) {
-		if($option_name === 'zuplus_clear_log') return $this->plugin->dbug->clear_log();
+		if($option_name === 'zuplus_clear_log') return $this->is_debug() ? $this->plugin->dbug->clear_log() : [];
 		if($option_name === 'zuplus_duplicate_menu') return zuplus_ajax_duplicate_menu();
+		if($option_name === 'zuplus_revoke_cookie') return ['info'	=> sprintf('Cookie "<strong>%1$s</strong>" was deleted', $ajax_value)];
 
 		if($option_name === 'zuplus_reset_cached') return $this->reset_cached();
 
@@ -365,6 +347,10 @@ function zuplus_get_my_url() {
 
 function zuplus_instance() { 
 	return ZU_Plugin::instance(); 
+}
+
+function zuplus_nodebug() { 
+	return !zuplus_instance()->is_debug(); 
 }
 
 function zuplus_options() {